feat(cni): expose main functions of cni and installer binary to be reusable
