+++ app/cni/pkg/cni/injector_darwin.go
@@ -1,4 +1,4 @@
-package main
+package cni
 
 import (
 	"errors"
