+++ app/cni/pkg/cni/injector_linux.go
@@ -1,4 +1,4 @@
-package main
+package cni
 
 import (
 	"io/ioutil"
