+++ app/cni/pkg/cni/injector_windows.go
@@ -1,4 +1,4 @@
-package main
+package cni
 
 import (
 	"errors"
