+++ app/cni/pkg/cni/suite_test.go
@@ -1,4 +1,4 @@
-package main_test
+package cni_test
 
 import (
 	"testing"
