+++ app/cni/pkg/install/consts.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 const kumaCniConfig = `{
 	"type": "kuma-cni",
