+++ app/cni/pkg/install/installer_config_test.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 import (
 	"io/ioutil"
