+++ app/cni/pkg/install/main_test.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 import (
 	. "github.com/onsi/ginkgo/v2"
