+++ app/cni/pkg/install/testdata/10-calico.conflist
