+++ app/cni/pkg/install/testdata/10-flannel-cni-injected.conf
