+++ app/cni/pkg/install/testdata/10-flannel.conf
