+++ app/cni/pkg/install/testdata/10-kuma-cni.conf
