+++ app/cni/pkg/install/testdata/prepare-chained-kuma-config/10-calico.conflist
