+++ app/cni/pkg/install/testdata/prepare-chained-kuma-config/token
