+++ app/cni/pkg/install/testdata/prepare-kubeconfig/ZZZ-kuma-cni-kubeconfig
