+++ app/cni/pkg/install/testdata/prepare-kubeconfig/ca.crt
