+++ app/cni/pkg/install/testdata/prepare-kubeconfig/token
