+++ app/cni/pkg/install/testdata/prepare-standalone-kuma-config/kuma-cni.conf.golden
