+++ app/cni/pkg/install/testdata/prepare-standalone-kuma-config/token
