+++ app/cni/pkg/install/validation.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 import (
 	"github.com/pkg/errors"
