+++ app/cni/pkg/install/validation_test.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 import (
 	"path"
